fun_projects/image_upload_api/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-multipart
boto3
Pillow